Refactor action setup
This commit moves the logic of RunConfig.DigestSetup to the action
package as action.InstallSetup. In addition the install action is
no longer an entity but just a couple of methods (InstallSetup and
InstallRun) that read and modifies the RunConfig struct.

Finally it adds in install command some logic to sanitize flags.

This way the install command can be expressed in three simple steps:

1. Basic checks on flags and config as part of 'cmd/install.go'
2. Setup install details from a rough RunConfig initiated and run
   all logic to determine partitions, images, etc. according to the
   install action specifics.
3. Run installation once installation details are ready

Signed-off-by: David Cassany <dcassany@suse.com>