Drop lsblk in favour of ghw library
 - ghw uses udev database and files under /sys and /proc to gather all
info
 - This uses that exclusively so no more lsblk
 - Probably means that the partition/disk identification utils can now
without root privs
 - Adds a new mock to mock a full disk(s) with partition(s) for tests
only
 - Makes tests use the new mock to mock the lsblk calls
 - Fixes a couple of leftovers on dir permissions

Unofrtunately this needs my own fork of ghw as upstream is missing a
patch to get the labels from the udev database, but it should be merged
upstream soon.

It also means that when running elemental from docker we need to link
the /run/udev/data volume, but this can be done on RO as we only read,
which would be a bit nicer. Commadn will look uglier thougth, but as
long as we explain why it needs that volume, it should all be good.

Signed-off-by: Itxaka <igarcia@suse.com>