Move out generic configuration from RunConfig
This commit is a first step to move generic configuration from
RunConfig to a new Config struct. In this commit, to narrow the scope
of changes the new Config struct is embedded into RunConfig, so
from a RunConfig PoV the change is transparent.

This allows to reuse Config struct in BuildConfig, so all the
interfaces initalizations can be reused there.

Signed-off-by: David Cassany <dcassany@suse.com>