Improve build-iso flags
This commit changes `--iso.lable` flag to `--label` and drops
`--iso.image` and `--iso.uefi` flags since this are unlikely to be used
and they can still be set by using a custom manifest.yaml.

Finally it also changes `--repo` flag behavior to append repos on top
of whatever is defined in the manifest or by default. Appended
repositories with `--repo` flag are always having highest priority.

Signed-off-by: David Cassany <dcassany@suse.com>