Try to get a decent luet tempdir
Currently luet defaults to the default tmpdir which is usually /tmp

This is a problem for machines with low memory during upgrades as /tmp
is a tmpfs of 25% of the pgysical memory and can be exhausted easily.

Instead allow overriding the luet default tmpDirBase and reuse the
getTempDir util to try and get the persistent partition if its exists
and its mounted.

This should not affect other commands as it first tries to use the
TMPDIR if exists and then use the partition only if mounted, falling back
to the default /tmp dir.

Signed-off-by: Itxaka <igarcia@suse.com>