Produce `state.yaml` file with installation data
This commit makes upgrade|reset|install to create and upgrade
`state.yaml` file including system wide data (deployed images,
partition labels, etc.)

This PR introduces the concept of installation state and stores such
a data in `state.yaml` in two different locations, state partition root
and recovery partition root.

The purpose of this duplication is to be able to always find the
state.yaml file in a known location regardless of the image we are
booting.

Signed-off-by: David Cassany <dcassany@suse.com>