Rework grub config and modules for install
Not sure where we were getting the grub.cfg before but it wasnt from the
install source, which is bad.

This patch changes to pick the grub from the install source and copy it
into the state partition.

It also copies the required grub modules into the state partition as we
no longer call the grub2-install which used to generate them. As we now
use the default grub.efi from the shim package, we no longer need to
generate the efi artifact and we only ship the 2 required extra modules

Signed-off-by: Itxaka <igarcia@suse.com>