Introduce ClearBootEntry for entry cleaning
Adds the ClearBootEntry functions which will clear any entries that
exists on install.

This is usually not an issue, but if you reinstall any elemental based
os several times on the same machine, you will end up with several
entries most of them pointing to broken entries.

This patch introduces a cleaning function that will go voer all the boot
entries and check if they match our entry name and clean them so we dont
have duplicates. This is usually a problem on install due to the
recreating of the EFI partition which means that its UUID changes, and
the entry, which uses that UUID to identify the partition with the efi
binary, is no longer valid.

Reset is not affected as reset only formats the partitions, does not
recreates them.

There is an alternative approach that uses the ShortFormPathFile whis is
just the file path with no disk information. I tested and it works but
it looks very problematic to me, as it could point easily to different
devices as it only provides a path.

Signed-off-by: Itxaka <igarcia@suse.com>