Also copy files to /EFI/elemental
EFI files are usually under the /EFI dir in its own directory, so
opensuse stores them in /EFI/opensuse, fedora under /EFI/fedora and so
on.

Its nice that we set the fallback so we always boot from elemental if
boot manager is broken, but lets be good citizens and store our
artifacts under elemental so they are contained and nothing can touch
them other than ourselves. Its also nice to point to the shim under the
proper directory and know that nothing can overwrite it as a side
effect.

Signed-off-by: Itxaka <igarcia@suse.com>