Fill persistent mountpoint on upgrade
Upgrade tries to mount the persistent partition during the upgrade
in order to use it as tmp dir to store and decompress images.
Unfortunately as the mountpoint is empty, this fails and it
always falls back to the /tmp dir which is a tmpfs dir with
25% of the RAM as space. This means that in a 4Gb device we get
a /tmp dir of 1Gb, which is not enough to download and decompress
a single image.

This used to work before so we probably introduced this regression
with some changes and never notice due our workers having 10Gb of RAM

This patch fills the persistent mountpoint on upgrade config load
so we can temp mount it during upgrade and maintains the same fallback

Signed-off-by: itxaka <igarcia@suse.com>