Create default grub config files on install/upgrade
By default we load some files in our grub.cfg from the state partiton in
order to allow derivatives to overrides several grub configs.

This files not existing is usually not an issue, just a small message
that grub failed to found them during boot, nothing else.

Unfortunately on some machines this can lead to grub hanging out for 30
minutes while trying to search for those files.

This patch fixes that by create any missing files during install and
upgrade with empty contents, so grub can find them instantly and avoids
the errors and the hang.

The files are formatted in the proper grub format so this does not
prevent from editing them afterwards.

The missing files func is done at the end of the rebrand in both install
and upgrade, by that time any files from the install/upgrade sources
should already be there, so it should not override anything custom that
its coming from the install/upgrade sources.

Signed-off-by: Itxaka <igarcia@suse.com>