Add post-* hooks
This commit adds post-* hooks where the after hooks were originally
defined. After hooks are moved to happen straight in sequence after the
after-*-chroot hooks.

The main motivation for such a change is having a hook that includes the
partition, the running system and the deployed image all mounted and
easily accessible.

Signed-off-by: David Cassany <dcassany@suse.com>