Elemental image deploy refactor
This commit refactors elemental image deploy procedure to always create
and prepare the root tree in a temporary folder before synching it to
a filesystem image most likely as a mounted loop device.

This change has a couple of immediate benefits:

1. We can precompute the root tree size before creating the filesystem
   image. Hence image sizes can be adjusted to root-tree size.

2. There is no path differentiation between filesystems. The root tree
   is prepared following the same logic independently of the target
   filesystem. Squashfs is no longer an exception, building the image
   follows the same logic as it was any other writable filesystem.

We also can argue this is a simple, robust and flexible logic compared
with the previous code.

The counter part is having to prepare the root-tree to later on copy it
to the final image, this causes the root-tree to be written twice, one
to prepare it and another one to sync it to the target location.

Signed-off-by: David Cassany <dcassany@suse.com>